Refactor some services
